+++ src/toolkit/matl_sell_policy.cc
@@ -180,7 +180,7 @@ std::set<BidPortfolio<Material>::Ptr> MatlSellPolicy::GetMatlBids(
   Request<Material>* req;
   Material::Ptr m, offer;
   double qty;
-  int n_full_bids;
+  int n_full_bids = 0;
   double bid_qty;  
   double remaining_qty;
   std::vector<double> bids;
@@ -199,6 +199,20 @@ std::set<BidPortfolio<Material>::Ptr> MatlSellPolicy::GetMatlBids(
       bid_qty = excl ? quantize_ : package_->GetFillMass(qty);
       if (bid_qty != 0) {
         n_full_bids = static_cast<int>(std::floor(qty / bid_qty));
+
+        // Throw if number of bids above limit or if casting to int caused
+        // overflow to negative int limit 
+        std::string s;
+        if (manager() != NULL)
+          s + " Agent: "
+            + Trader::manager()->prototype() + "-"
+            + std::to_string(Trader::manager()->id()) + ". ";
+        s + "This is likely due to too much material (did you forget a "
+          + "throughput?) or small package limits relative to the "
+          + "quantity available. qty: " + std::to_string(qty) 
+          + ", and each bid would be: " + std::to_string(bid_qty);
+        Package::ExceedsSplitLimits(n_full_bids, s);
+
         bids.assign(n_full_bids, bid_qty);
 
         remaining_qty = fmod(qty, bid_qty);