+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ Since last release
 ====================
 
 **Added:**
+
 * Added TransportUnits (#1750, #1772)
+* Added CI support for Ubuntu 24.04 (#1770)
 
 **Changed:**